Renamed things a bit
