--- aqserver/go.mod
@@ -1,8 +1,0 @@
-module github.com/ExploratoryEngineering/air-quality-sensor-node/aqserver
-
-go 1.13
-
-require (
-	github.com/stretchr/testify v1.5.1
-	github.com/telenordigital/nbiot-go v0.0.0-20200302123245-033df8c1f27d
-)
