+++ server/go.mod
@@ -1,0 +1,5 @@
+module github.com/ExploratoryEngineering/air-quality-sensor-node/server
+
+go 1.13
+
+require github.com/telenordigital/nbiot-go v0.0.0-20200302123245-033df8c1f27d
